Pass list of layers via plot_params. Refactor overlay_pass fn
